+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - animating errors show / hide now
   - fixed radio/checkboxes data-trigger behavior (#115)
   - added `change` event to quickly remove radio/checkboxes errors (#115)
+  - added ability to group radio / checkboxes by `data-group` attribute intead of name
 
 ** 1.1.9 **